requirements.txt
discord.py[voice]
youtube_dl==2020.9.20
textblob